pyproject.toml
[tool.poetry]
name = "doc-extraction"
version = "0.1.0"
description = "Document extraction for PDFs, DOCX, images, and tables using Apache Tika, pdfplumber, python-docx, and Camelot."
authors = ["User <user@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
tika = "*"
pdfplumber = "*"
python-docx = "*"
camelot-py = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
